+++ packages/imageeditor/src/index.js
@@ -112,14 +112,12 @@ class ImageEditor extends PureComponent {
         }
     }
 
-    /**
-     * When maintaining ref with inline arrow function, it sometimes receives
-     * mystery `null` during canvas drag. This blocks accessing cropping rect
-     * via `onImageChange()` callback.
-     *
-     * Setting ref with instance method doesn't break. It's just MAGIC.
-     */
-    getImageCanvas = () => this.editorRef.current.getImage();
+    getImageCanvas = ({ originalSize = false } = {}) => {
+        if (originalSize) {
+            return this.editorRef.current.getImage();
+        }
+        return this.editorRef.current.getImageScaledToCanvas();
+    }
 
     handleSliderChange = (event) => {
         const newScale = Number(event.target.value);